feat: Add the ability to generate and validate server configs from the CLI
