+++ src/aux-records/index.ts
@@ -55,3 +55,4 @@ export * from './websockets';
 export * from './forms';
 
 export * from './ComIdConfig';
+export * from './ServerConfig';
